Added getSelectedToken util token
