+++ src/components/common/Dialogs/CreatePaymentDialog/TokenAmountInput/TokenAmountInput.tsx
@@ -5,7 +5,7 @@ import { useFormContext } from 'react-hook-form';
 
 import { HookFormInput as Input, TokenSymbolSelector } from '~shared/Fields';
 import EthUsd from '~shared/EthUsd';
-import { getTokenDecimalsWithFallback } from '~utils/tokens';
+import { getSelectedToken, getTokenDecimalsWithFallback } from '~utils/tokens';
 import { notNull } from '~utils/arrays';
 import { Colony } from '~types';
 
@@ -32,20 +32,18 @@ const MSG = defineMessages({
 });
 
 interface Props {
-  colony: Colony | undefined;
+  colony: Colony;
   disabled: boolean;
 }
 
 const TokenAmountInput = ({ colony, disabled }: Props) => {
   const { getValues } = useFormContext();
   const values = getValues();
+  const selectedToken = getSelectedToken(colony, values.tokenAddress);
   const colonyTokens =
     colony?.tokens?.items
       .filter(notNull)
       .map((colonyToken) => colonyToken.token) || [];
-  const selectedToken = colonyTokens.find(
-    (token) => token?.tokenAddress === values.tokenAddress,
-  );
   const formattingOptions = useMemo(
     () => ({
       delimiter: ',',
