+++ src/utils/tokens.ts
@@ -3,6 +3,9 @@ import { BigNumber, BigNumberish } from 'ethers';
 import moveDecimal from 'move-decimal-point';
 
 import { DEFAULT_TOKEN_DECIMALS } from '~constants';
+import { Colony } from '~types';
+
+import { notNull } from './arrays';
 
 export const getBalanceFromToken = (
   /** @TODO: add proper type */
@@ -84,4 +87,15 @@ export const calculateFee = (
     feesInWei: feesInWei.toString(),
     totalToPay: moveDecimal(totalToPayInWei, -1 * decimals),
   }; // NOTE: seems like moveDecimal does not have strict typing
+};
+
+export const getSelectedToken = (colony: Colony, tokenAddress: string) => {
+  const colonyTokens =
+    colony?.tokens?.items
+      .filter(notNull)
+      .map((colonyToken) => colonyToken.token) || [];
+  const selectedToken = colonyTokens.find(
+    (token) => token?.tokenAddress === tokenAddress,
+  );
+  return selectedToken;
 };
